+++ lib/grpc_generator.dart
@@ -96,12 +96,11 @@ class GrpcServiceGenerator {
       // TODO(jakobr): Throw more actionable error.
       throw 'FAILURE: Unknown type reference (${fqname}) for ${location}';
     }
-    if (fileGen.package == mg.fileGen.package || mg.fileGen.package == '') {
-      // It's either the same file, or another file with the same package.
-      // (In the second case, we import it without using "as".)
+    if (fileGen.protoFileUri == mg.fileGen.protoFileUri) {
+      // If it's the same file, we import it without using "as".
       return mg.classname;
     }
-    return mg.packageImportPrefix + '.' + mg.classname;
+    return mg.fileImportPrefix + '.' + mg.classname;
   }
 
   void generate(IndentingWriter out) {